+++ packages/nimble-components/src/table/template.ts
@@ -1,9 +1,27 @@
-import { html, repeat } from '@microsoft/fast-element';
+import {
+    ElementsFilter,
+    html,
+    repeat,
+    slotted,
+    when
+} from '@microsoft/fast-element';
 import { DesignSystem } from '@microsoft/fast-foundation';
 import type { Table } from '.';
 import type { TableRowState } from './types';
 import { TableHeader } from './components/header';
 import { TableRow } from './components/row';
+import { TableColumn } from '../table-column/base';
+
+const isTableColumn = (): ElementsFilter => {
+    const filter: ElementsFilter = (
+        value: Node,
+        _: number,
+        __: Node[]
+    ): boolean => {
+        return value instanceof TableColumn;
+    };
+    return filter;
+};
 
 // prettier-ignore
 export const template = html<Table>`
@@ -11,22 +29,27 @@ export const template = html<Table>`
         <div class="table-container">
             <div role="rowgroup" class="header-container">
                 <div class="header-row" role="row">
-                    ${repeat(x => x.columnHeaders, html<string>`
+                    ${repeat(x => x.columns, html<TableColumn>`
                         <${DesignSystem.tagFor(TableHeader)} class="header">
-                            ${x => x}
+                            ${x => x.textContent}
                         </${DesignSystem.tagFor(TableHeader)}>
                     `)}
                 </div>
             </div>
             <div class="table-viewport" role="rowgroup">
+            ${when(x => x.columns.length > 0, html<Table>`
                 ${repeat(x => x.tableData, html<TableRowState>`
                     <${DesignSystem.tagFor(TableRow)}
-                        :data="${x => x.data}"
+                        class="row"
+                        record-id="${x => x.id}"
+                        :dataRecord="${x => x.record}"
                         :columns="${(_, c) => (c.parent as Table).columns}"
                     >
                     </${DesignSystem.tagFor(TableRow)}>
                 `)}
+            `)}
             </div>
         </div>
+        <slot ${slotted({ property: 'columns', filter: isTableColumn() })}></slot>
     </template>
 `;